move api calls from details to registeration
moved all the api calls from Player Details to Account Registration so that when the user makes an account the api loads there. and when details is accessed, it uses the DB instead of api calls.
However it is not being saved in the database yet. That will be the next step